Add language combo box to monitor
